+++ src/hooks/usePlanCandidateGalleryPageAutoScroll.ts
