+++ src/view/navigation/NavBar.tsx
@@ -3,13 +3,13 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import { ReactNode } from "react";
 import { MdArrowBack } from "react-icons/md";
-import { reduxHistorySelector } from "src/redux/history";
-import AppLogoImage from "src/view/assets/svg/app_logo_horizontal.svg";
 import { Padding } from "src/constant/padding";
 import { Routes } from "src/constant/router";
 import { Size } from "src/constant/size";
-import { useAuth } from "src/view/hooks/useAuth";
-import { useBindPreLoginState } from "src/view/hooks/useBindPreLoginState";
+import { useAuth } from "src/hooks/useAuth";
+import { useBindPreLoginState } from "src/hooks/useBindPreLoginState";
+import { reduxHistorySelector } from "src/redux/history";
+import AppLogoImage from "src/view/assets/svg/app_logo_horizontal.svg";
 import { BindPreLoginStateConfirmationDialog } from "src/view/plandetail/BindPreLoginStateConfirmationDialog";
 import { NavBarUser } from "src/view/user/NavBarUser";
 import styled from "styled-components";
