+++ src/view/plancandidate/DialogAddPlace.tsx
@@ -1,7 +1,7 @@
 import { Place } from "src/domain/models/Place";
 import { PlacesWithCategory } from "src/domain/models/PlacesWithCategory";
 import { Transition } from "src/domain/models/Transition";
-import { useAppTranslation } from "src/view/hooks/useAppTranslation";
+import { useAppTranslation } from "src/hooks/useAppTranslation";
 import { DialogRelatedPlaces } from "src/view/plancandidate/DialogRelatedPlaces";
 
 type Props = {
