+++ src/view/plandetail/PlaceChipContextAction.tsx
@@ -11,8 +11,8 @@ import {
 } from "react-icons/md";
 import { SiGooglemaps, SiInstagram } from "react-icons/si";
 import { AnalyticsEvents } from "src/constant/analytics";
-import { useAppTranslation } from "src/view/hooks/useAppTranslation";
-import { UploadPlaceImageProps } from "src/view/hooks/useUploadPlaceImage";
+import { useAppTranslation } from "src/hooks/useAppTranslation";
+import { UploadPlaceImageProps } from "src/hooks/useUploadPlaceImage";
 import { OnClickHandler } from "src/view/types/handler";
 
 type Props = {
