+++ src/view/plandetail/PlaceInfoTab.tsx
@@ -3,11 +3,11 @@ import { useTranslation } from "next-i18next";
 import { ReactNode, useState } from "react";
 import { IconType } from "react-icons";
 import { MdCurrencyYen, MdSchedule } from "react-icons/md";
+import { Size } from "src/constant/size";
 import { PlaceCategory } from "src/domain/models/PlaceCategory";
 import { PriceRange } from "src/domain/models/PriceRange";
 import { DateHelper } from "src/domain/util/date";
-import { Size } from "src/constant/size";
-import { useAppTranslation } from "src/view/hooks/useAppTranslation";
+import { useAppTranslation } from "src/hooks/useAppTranslation";
 import { getPlaceCategoryIcon } from "src/view/plan/PlaceCategoryIcon";
 
 type Props = {
